// types/dom.d.ts
export interface AppElements {
  movieList: HTMLUListElement;
  searchForm: HTMLFormElement;
  searchInput: HTMLInputElement;

  heroBackdrop: HTMLDivElement;
  heroRate: HTMLDivElement;
  heroRateValue: HTMLSpanElement;
  heroTitle: HTMLHeadingElement;

  skeletonCard: HTMLUListElement;
  seeMoreBtn: HTMLButtonElement;
}